Rename prisma-e2e-windows-test to prisma-e2e-windows-test-new + target es2018
