+++ platforms/azure-functions-windows/prisma-e2e-windows-test-new/function.json
@@ -1,7 +1,6 @@
 {
   "disabled": false,
-  "bindings": [
-    {
+  "bindings": [{
       "authLevel": "function",
       "type": "httpTrigger",
       "direction": "in",
