+++ platforms/azure-functions-windows/run.sh
@@ -15,6 +15,6 @@ yarn install
 yarn prisma2 generate
 yarn tsc
 
-func azure functionapp publish "prisma-e2e-windows-test" --force
+func azure functionapp publish "prisma-e2e-windows-test-new" --force
 
 sh test.sh
