+++ platforms/azure-functions-windows/tsconfig.json
@@ -1,11 +1,12 @@
 {
   "exclude": [],
   "include": [
-		"prisma-e2e-windows-test/"
+    "prisma-e2e-windows-test-new/"
   ],
   "compilerOptions": {
     "sourceMap": true,
-    "outDir": "prisma-e2e-windows-test/",
+    "outDir": "prisma-e2e-windows-test-new/",
+    "target": "es2018",
     "strict": true,
     "lib": ["esnext", "dom"],
     "esModuleInterop": true
